requirements.txt
# Core Flask dependencies
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.3
flask-cors==5.0.1
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==7.4.2
python-dotenv==1.0.0
SQLAlchemy>=1.4.0
typing_extensions==4.12.2
Werkzeug==2.3.7

# Task scheduling
APScheduler==3.11.0

# Production server
gunicorn==21.2.0

# Database drivers
pg8000==1.29.8
psycopg2-binary==2.9.9  # Required for PostgreSQL

# Development tools
pytest-cov==4.1.0  # For test coverage
black==24.2.0  # For code formatting
flake8==7.0.0  # For code linting

# Observability and monitoring
sentry-sdk[flask]==1.40.5  # Error tracking
prometheus-flask-exporter==0.23.0  # Metrics

# Security
bcrypt==4.1.2  # Password hashing
cryptography==42.0.5  # Encryption utilities

# API utilities
marshmallow==3.20.2  # Object serialization/deserialization
webargs==8.4.0  # Request parsing